Revert "Logout"
