--- .eslintignore
@@ -1,2 +1,0 @@
-node_modules
-src/modules/_global/scrollableTabView
