--- .eslintrc
@@ -1,109 +1,0 @@
-{
-	"extends": [
-		"eslint:recommended",
-		"plugin:import/errors",
-		"plugin:import/warnings",
-		"airbnb"
-	],
-	"plugins": [
-		"react",
-		"react-native"
-	],
-	"parserOptions": {
-		"ecmaVersion": 6,
-		"sourceType": "module",
-		"ecmaFeatures": {
-			"jsx": true
-		}
-	},
-	"env": {
-		"browser": true,
-		"es6": true,
-		"jquery": true,
-		"mocha": true,
-		"node": true
-	},
-	"rules": {
-		"arrow-body-style": [1, "as-needed"],
-		"arrow-parens": [1, "as-needed"],
-		"class-methods-use-this": 0,
-		"comma-dangle": [1, "never"],
-		"consistent-return": 0,
-		"eol-last": 0,
-		"func-names": 0,
-		"import/extensions": 0,
-		"import/imports-first": 0,
-		"import/no-extraneous-dependencies": 0,
-		"import/no-unresolved": 0,
-		"indent": ["error", 4, { "SwitchCase": 1, "VariableDeclarator": 1 }],
-		"jsx-a11y/href-no-hash": 0,
-		"jsx-quotes": 1,
-		"linebreak-style": 0,
-		"max-len": 0,
-        "new-cap": 0,
-		"no-alert": 0,
-		"no-console": 1,
-		"no-debugger": 1,
-		"no-global-assign": 0,
-		"no-lone-blocks": 0,
-		"no-multi-spaces": 1,
-		"no-shadow": 0,
-		"no-tabs": 0,
-		"no-trailing-spaces": 1,
-		"no-underscore-dangle": 0,
-		"no-unsafe-negation": 0,
-		"no-unused-vars": [1, { "varsIgnorePattern": "(bindActionCreators|colors|chalk|dotenv|PropTypes)", "args": "none" }],
-		"no-use-before-define": 0,
-		"no-useless-constructor": 0,
-		"no-var": 1,
-		"object-curly-spacing": 1,
-		"one-var": 0,
-		"quotes": 0,
-		"react-native/no-color-literals": 0,
-		"react-native/no-inline-styles": 0,
-		"react-native/no-unused-styles": 2,
-		"react-native/split-platform-components": 2,
-		"react/display-name": [ 1, { "ignoreTranspilerName": false }],
-		"react/forbid-prop-types": [1, { "forbid": ["any"] }],
-		"react/jsx-boolean-value": 1,
-		"react/jsx-closing-bracket-location": [1, { "selfClosing": "tag-aligned", "nonEmpty": "after-props" }],
-		"react/jsx-curly-spacing": 1,
-		"react/jsx-filename-extension": 0,
-		"react/jsx-indent": ["error", 4],
-		"react/jsx-indent-props": 0,
-		"react/jsx-key": 1,
-		"react/jsx-max-props-per-line": [1, { "maximum": 5 }],
-		"react/jsx-no-bind": 0,
-		"react/jsx-no-duplicate-props": 1,
-		"react/jsx-no-literals": 0,
-		"react/jsx-no-undef": 1,
-		"react/jsx-pascal-case": 1,
-		"react/jsx-sort-prop-types": 0,
-		"react/jsx-sort-props": 0,
-		"react/jsx-space-before-closing": [1, "always"],
-		"react/jsx-uses-react": 1,
-		"react/jsx-uses-vars": 1,
-		"react/jsx-wrap-multilines": 1,
-		"react/no-danger": 1,
-		"react/no-did-mount-set-state": 1,
-		"react/no-did-update-set-state": 1,
-		"react/no-direct-mutation-state": 1,
-		"react/no-multi-comp": 1,
-		"react/no-set-state": 0,
-		"react/no-unknown-property": 1,
-		"react/prefer-es6-class": 1,
-		"react/prefer-stateless-function": 0,
-		"react/prop-types": 1,
-		"react/react-in-jsx-scope": 1,
-		"react/require-extension": 0,
-		"react/self-closing-comp": 1,
-		"react/sort-comp": 1,
-		"semi": [1, "always"]
-	},
-	"settings": {
-		"react": {
-			"pragma": "React",
-			"version": "15.3.2"
-		}
-	}
-}
