+++ App.js
@@ -1,0 +1,3 @@
+import AppNavigator from "./app/config/routes";
+
+export default AppNavigator;
