+++ android/app/src/main/java/com/travelapp/MainActivity.java
@@ -1,7 +1,15 @@
 package com.travelapp;
 
-import com.reactnativenavigation.controllers.SplashActivity;
+import com.facebook.react.ReactActivity;
 
-public class MainActivity extends SplashActivity {
+public class MainActivity extends ReactActivity {
 
+    /**
+     * Returns the name of the main component registered from JavaScript.
+     * This is used to schedule rendering of the component.
+     */
+    @Override
+    protected String getMainComponentName() {
+        return "TravelApp";
+    }
 }
