+++ android/app/src/main/java/com/travelapp/MainApplication.java
@@ -1,43 +1,45 @@
 package com.travelapp;
 
 import android.app.Application;
-import android.util.Log;
-import android.support.annotation.NonNull;
 
 import com.facebook.react.ReactApplication;
-import com.facebook.react.ReactInstanceManager;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
-import com.reactnativenavigation.NavigationApplication;
 import com.facebook.soloader.SoLoader;
 
-import java.util.List;
 import java.util.Arrays;
+import java.util.List;
 
-public class MainApplication extends NavigationApplication {
+public class MainApplication extends Application implements ReactApplication {
 
-     @Override
-     public boolean isDebug() {
-         // Make sure you are using BuildConfig from your own application
-         return BuildConfig.DEBUG;
-     }
+  private final ReactNativeHost mReactNativeHost = new ReactNativeHost(this) {
+    @Override
+    public boolean getUseDeveloperSupport() {
+      return BuildConfig.DEBUG;
+    }
 
-     protected List<ReactPackage> getPackages() {
-         // Add additional packages you require here
-         // No need to add RnnPackage and MainReactPackage
-         return Arrays.<ReactPackage>asList(
-             // eg. new VectorIconsPackage()
-         );
-     }
+    @Override
+    protected List<ReactPackage> getPackages() {
+      return Arrays.<ReactPackage>asList(
+          new MainReactPackage()
+      );
+    }
 
-     @Override
-     public List<ReactPackage> createAdditionalReactPackages() {
-         return getPackages();
-     }
+    @Override
+    protected String getJSMainModuleName() {
+      return "index";
+    }
+  };
 
-     @Override
-public String getJSMainModuleName() {
-    return "index";
+  @Override
+  public ReactNativeHost getReactNativeHost() {
+    return mReactNativeHost;
+  }
+
+  @Override
+  public void onCreate() {
+    super.onCreate();
+    SoLoader.init(this, /* native exopackage */ false);
+  }
 }
- }
