+++ android/settings.gradle
@@ -1,6 +1,3 @@
-rootProject.name = 'travelapp'
+rootProject.name = 'TravelApp'
 
 include ':app'
-
-include ':react-native-navigation'
-project(':react-native-navigation').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-navigation/android/app/')
