+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "travel-app",
+  "name": "TravelApp",
   "displayName": "TravelApp"
 }
