--- app/constants/actionTypes.js
@@ -1,2 +1,0 @@
-export const PERSIST_LOGIN = 'persist_login';
-export const LOGOUT = 'logout';
