--- app/constants/colors.js
@@ -1,7 +1,0 @@
-const colors = {
-    primaryColor: '#518ffb',
-    primaryTextColor: '#000',
-    dimmedTextColor: '#D0D0D0'
-};
-
-export default colors;
