--- app/modules/home/home.actions.js
@@ -1,6 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-import * as actionTypes from '../../constants/actionTypes';
-
-export const logout = () => ({
-    type: actionTypes.LOGOUT
-});
