--- app/modules/register/index.js
@@ -1,4 +1,0 @@
-import { connect } from "react-redux";
-import RegisterPage from "./register.component";
-
-export default connect(null, null)(RegisterPage);
