+++ package.json
@@ -1,13 +1,11 @@
 {
-  "name": "travel-app",
+  "name": "TravelApp",
   "version": "0.0.1",
   "private": true,
   "scripts": {
     "android": "react-native run-android",
     "ios": "react-native run-ios",
-    "link": "react-native link",
-    "lint": "esw index.js src utils",
-    "lint:watch": "npm run lint -- --watch"
+    "link": "react-native link"
   },
   "rnpm": {
     "assets": [
@@ -15,29 +13,19 @@
     ]
   },
   "dependencies": {
-    "eslint": "^3.9.1",
-    "eslint-config-airbnb": "^13.0.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "^6.6.0",
-    "eslint-plugin-react-native": "^2.0.0",
-    "eslint-watch": "2.1.14",
     "prop-types": "latest",
     "react": "16.3.1",
     "react-native": "0.55.3",
+    "react-native-check-box": "^2.1.0",
     "react-native-custom-checkbox": "^1.5.2",
-    "react-native-navigation": "latest",
-    "react-redux": "^5.0.5",
-    "redux": "^3.6.0",
-    "redux-thunk": "^2.2.0"
+    "react-navigation": "^1.5.11"
   },
   "devDependencies": {
     "babel-jest": "22.4.3",
     "babel-preset-react-native": "4.0.0",
+    "eslint": "^4.19.1",
     "jest": "22.4.3",
-    "react-test-renderer": "16.3.1",
-    "redux-immutable-state-invariant": "^2.0.0",
-    "redux-logger": "^3.0.6"
+    "react-test-renderer": "16.3.1"
   },
   "jest": {
     "preset": "react-native"
